+++ t/app/controller/admin/sitemessage.t
@@ -25,35 +25,35 @@ $ukc->mock('_get_bank_holiday_json', sub {
 FixMyStreet::override_config {
     ALLOWED_COBRANDS => [ 'oxfordshire' ],
 }, sub {
-    subtest 'setting emergency message' => sub {
+    subtest 'setting site message' => sub {
         $user->user_body_permissions->create({
             body => $body,
             permission_type => 'emergency_message_edit',
         });
 
-        $mech->get_ok('/admin/emergencymessage');
+        $mech->get_ok('/admin/sitemessage');
         $mech->content_lacks('Waste message');
-        $mech->submit_form_ok({ with_fields => { emergency_message => 'Testing emergency message' } });
-        $mech->content_contains('Testing emergency message');
+        $mech->submit_form_ok({ with_fields => { site_message => 'Testing site message' } });
+        $mech->content_contains('Testing site message');
         $mech->get_ok('/');
-        $mech->content_contains('Testing emergency message');
+        $mech->content_contains('Testing site message');
 
         # Check removing message
-        $mech->get_ok('/admin/emergencymessage');
-        $mech->submit_form_ok({ with_fields => { emergency_message => '' } });
-        $mech->content_lacks('Testing emergency message');
+        $mech->get_ok('/admin/sitemessage');
+        $mech->submit_form_ok({ with_fields => { site_message => '' } });
+        $mech->content_lacks('Testing site message');
         $mech->get_ok('/');
-        $mech->content_lacks('Testing emergency message');
+        $mech->content_lacks('Testing site message');
     };
 
-    subtest "user without permissions can't set emergency message" => sub {
+    subtest "user without permissions can't set site message" => sub {
         $user->user_body_permissions->delete;
         $user->user_body_permissions->create({
             body => $body,
             permission_type => 'report_edit',
         });
 
-        $mech->get('/admin/emergencymessage');
+        $mech->get('/admin/sitemessage');
         ok !$mech->res->is_success, "want a bad response";
         is $mech->res->code, 404, "got 404";
     };
@@ -63,27 +63,27 @@ FixMyStreet::override_config {
     ALLOWED_COBRANDS => [ 'oxfordshire' ],
     COBRAND_FEATURES => { waste => { oxfordshire => 1 } },
 }, sub {
-    subtest 'setting emergency waste message' => sub {
+    subtest 'setting site waste message' => sub {
         $user->user_body_permissions->create({
             body => $body,
             permission_type => 'emergency_message_edit',
         });
 
-        $mech->get_ok('/admin/emergencymessage');
+        $mech->get_ok('/admin/sitemessage');
         $mech->content_contains('Waste message');
-        $mech->submit_form_ok({ with_fields => { emergency_message_waste => 'Testing emergency waste message' } });
-        $mech->content_contains('Testing emergency waste message');
+        $mech->submit_form_ok({ with_fields => { site_message_waste => 'Testing site waste message' } });
+        $mech->content_contains('Testing site waste message');
         $mech->get_ok('/');
-        $mech->content_lacks('Testing emergency waste message');
+        $mech->content_lacks('Testing site waste message');
         $mech->get_ok('/waste');
-        $mech->content_contains('Testing emergency waste message');
+        $mech->content_contains('Testing site waste message');
 
         # Check removing message
-        $mech->get_ok('/admin/emergencymessage');
-        $mech->submit_form_ok({ with_fields => { emergency_message_waste => '' } });
-        $mech->content_lacks('Testing emergency waste message');
+        $mech->get_ok('/admin/sitemessage');
+        $mech->submit_form_ok({ with_fields => { site_message_waste => '' } });
+        $mech->content_lacks('Testing site waste message');
         $mech->get_ok('/waste');
-        $mech->content_lacks('Testing emergency waste message');
+        $mech->content_lacks('Testing site waste message');
     };
 };
 
@@ -91,7 +91,7 @@ FixMyStreet::override_config {
     ALLOWED_COBRANDS => [ 'oxfordshire' ],
 }, sub {
     subtest 'setting reporting message' => sub {
-        $mech->get_ok('/admin/emergencymessage');
+        $mech->get_ok('/admin/sitemessage');
         $mech->content_contains('hard-coded');
     }
 };
@@ -108,8 +108,8 @@ FixMyStreet::override_config {
         });
 
         $mech->host('bexley.example.org');
-        $mech->get_ok('/admin/emergencymessage');
-        $mech->submit_form_ok({ with_fields => { emergency_message_reporting => 'Testing reporting message' } });
+        $mech->get_ok('/admin/sitemessage');
+        $mech->submit_form_ok({ with_fields => { site_message_reporting => 'Testing reporting message' } });
         $mech->content_contains('Testing reporting message');
         $mech->get_ok('/report/new?latitude=51.45556&longitude=0.15356');
         $mech->content_contains('Testing reporting message');
@@ -120,8 +120,8 @@ FixMyStreet::override_config {
 
         # Check removing message
         $mech->host('bexley.example.org');
-        $mech->get_ok('/admin/emergencymessage');
-        $mech->submit_form_ok({ with_fields => { emergency_message_reporting => '' } });
+        $mech->get_ok('/admin/sitemessage');
+        $mech->submit_form_ok({ with_fields => { site_message_reporting => '' } });
         $mech->content_lacks('Testing reporting message');
         $mech->get_ok('/report/new?latitude=51.45556&longitude=0.15356');
         $mech->content_lacks('Testing reporting message');
@@ -129,9 +129,9 @@ FixMyStreet::override_config {
 
     subtest 'setting OOH messages' => sub {
         my ($csrf) = $mech->content =~ /name="token" value="([^"]*)"/;
-        $mech->post_ok('/admin/emergencymessage', {
-            emergency_message => 'Testing message',
-            emergency_message_ooh => 'This is an OOH message',
+        $mech->post_ok('/admin/sitemessage', {
+            site_message => 'Testing message',
+            site_message_ooh => 'This is an OOH message',
             # Tuesdays, midnight-noon
             'ooh[0].day' => 3,
             'ooh[0].start' => 0,